refactor(send): pull some fns into methods of Pos
