+++ lua/jet/core/send/get_range.lua
@@ -23,7 +23,7 @@ end
 ---@return jet.send.Range?
 M.get_expr = function(p)
 	-- Note: we want the filetype at the _cursor_, not the buffer filetype
-	local ft = require("jet.core.send.utils").local_lang_info(p).filetype
+	local ft = p:lang_info().filetype
 	local ft_module = M.filetype[ft]
 	---@diagnostic disable-next-line: unnecessary-if
 	if ft_module and ft_module.get_expr then
@@ -125,11 +125,7 @@ M._handle_curr_motion = function(mode)
 		end_col = pos2[3],
 	}
 
-	local ft = require("jet.core.send.utils").local_lang_info(pos.new({
-		buf = code.buf,
-		row = code.start_row,
-		col = code.start_col,
-	})).filetype
+	local ft = pos.new({ buf = code.buf, row = code.start_row, col = code.start_col }):lang_info().filetype
 
 	---`if` to avoid LSP warnings
 	if _G.JET_OP_PENDING_CALLBACK then
