+++ lua/jet/core/send/pos.lua
@@ -82,4 +82,74 @@ function Pos:to_char()
 	return line:sub(self.col + 1, self.col + 1)
 end
 
+---Adapted from https://github.com/neovim/neovim/blob/master/runtime/lua/vim/_comment.lua
+---NOTE: if this causes issues in the future (e.g. we don't actually want the
+---range-specific filetype) we could instead return a table of candidate
+---filetypes and commentstrings, then match these against kernel filetypes
+---until we find a match. Don't do this until it's clear that it's a real issue.
+---@return { filetype?: string, commentstring?: string }
+function Pos:lang_info()
+	self = self or Pos.get_curr()
+	local buf_ft = vim.bo[self.buf].filetype
+	local buf_cs = vim.bo[self.buf].commentstring
+
+	local ts_parser = vim.treesitter.get_parser(self.buf, "")
+	if not ts_parser then
+		return {
+			filetype = buf_ft,
+			commentstring = buf_cs,
+		}
+	end
+
+	-- Get 'commentstring' from tree-sitter captures' metadata.
+	-- Traverse backwards to prefer narrower captures.
+	local captures = vim.treesitter.get_captures_at_pos(self.buf, self.row, self.col - 1)
+	for i = #captures, 1, -1 do
+		local id, metadata = captures[i].id, captures[i].metadata
+		local metadata_cs = metadata["bo.commentstring"] or metadata[id] and metadata[id]["bo.commentstring"]
+		local metadata_ft = metadata["bo.filetype"] or metadata[id] and metadata[id]["bo.filetype"]
+
+		if metadata_cs and metadata_ft and type(metadata_ft) == "string" then
+			return {
+				filetype = metadata_ft,
+				commentstring = metadata_cs,
+			}
+		end
+	end
+
+	-- Get filetype and commentstring from the deepest LanguageTree which
+	-- both contains reference range and has valid 'commentstring' and
+	-- 'filetype'. In simple cases using `parser:language_for_range()` would be
+	-- enough, but it might return a language without a valid 'commentstring',
+	-- (like 'comment'), which is not very useful for our purposes.
+	local treesitter_ft, treesitter_cs, res_level = nil, nil, 0
+
+	---@param lang_tree vim.treesitter.LanguageTree
+	local function traverse(lang_tree, level)
+		if not lang_tree:contains({ self.row, self.col, self.row, self.col }) then
+			return
+		end
+
+		local filetypes = vim.treesitter.language.get_filetypes(lang_tree:lang())
+		for _, curr_ft in ipairs(filetypes) do
+			local cur_cs = vim.filetype.get_option(curr_ft, "commentstring")
+			if cur_cs ~= "" and curr_ft ~= "" and level > res_level then
+				treesitter_cs = cur_cs
+				treesitter_ft = curr_ft
+				break
+			end
+		end
+
+		for _, child_lang_tree in pairs(lang_tree:children()) do
+			traverse(child_lang_tree, level + 1)
+		end
+	end
+	traverse(ts_parser, 1)
+
+	return {
+		filetype = (treesitter_ft or buf_ft),
+		commentstring = (treesitter_cs or buf_cs),
+	}
+end
+
 return Pos
